QR Codes Using Flask/qr_app_pt1/requirements.txt
Click==7.0
euclid3==0.1
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.15.4
prettytable==0.7.2
pypng==0.0.18
PyQRCode==1.2.1
solidpython==0.2.0
Werkzeug==0.14.1
